+++ src/inspect_ai/approval/__init__.py
@@ -1,7 +1,7 @@
 from ._approval import Approval, ApprovalDecision
 from ._approver import Approver
 from ._auto import auto_approver
-from ._human import human_approver
+from ._human.approver import human_approver
 from ._policy import ApprovalPolicy
 from ._registry import approver
